// src/im/types.ts
export interface ConversationAddress {
  conversationId: string
  conversationShortId: string
  conversationType: 1 | 2
  inboxType?: number
}